# docs/swagger.yaml
basePath: /api/v1
definitions:
  middleware.Login:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  model.Todo:
    properties:
      _id:
        type: string
      completed:
        type: boolean
      created_at:
        type: string
      text:
        type: string
      updated_at:
        type: string
    type: object
  model.TodoDocInput:
    properties:
      completed:
        type: boolean
      text:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: pattercm@gmail.com
    name: Charles Patterson
    url: https://github.com/CharlesPatterson/
  description: CLI and API for managing TODOs in MongoDB
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Gin Todo API
  version: "1.0"
paths:
  /login:
    post:
      operationId: login
      parameters:
      - description: Login credentials
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/middleware.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Todo'
      summary: Login
      tags:
      - Auth
  /todos:
    get:
      description: Get all todos without any filtering
      operationId: get-all-todos
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Todo'
            type: array
      security:
      - JWT: []
      summary: Get all todos
      tags:
      - Todos
    post:
      operationId: create-todo
      parameters:
      - description: Todo data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.TodoDocInput'
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Todo'
      security:
      - JWT: []
      summary: Create a todo
      tags:
      - Todos
  /todos/{id}:
    delete:
      operationId: delete-todo-by-id
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.Todo'
      security:
      - JWT: []
      summary: Delete a todo
      tags:
      - Todos
    get:
      operationId: get-todo-by-id
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Todo'
      security:
      - JWT: []
      summary: Get a TODO by ID
      tags:
      - Todos
    put:
      operationId: update-todo-by-id
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      - description: Todo data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.TodoDocInput'
      - description: Authorization
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.Todo'
      security:
      - JWT: []
      summary: Update a TODO by ID
      tags:
      - Todos
schemes:
- http
- https
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"